:bangbang: extract supabase and most of the swr stuff out to services modules! :feelsgood:
this is pretty huge, so setup a common sorta-cart-services repo that will be a git submodule in src/services/ dir and house abstractions related to getting remote (or local) data. this also re-works a bunch of the types, it got a little kludge-y at points, so will hopefully revisit that later.